Taking an approach to dynamically determining FedEx API version
This feels a little obtuse though it does the job of injecting all the
XML namespaces correctly so that there shouldn't be a bunch of rework
done on the rest of the codebase.

Most of the tests are going to fail. This made the following test pass

FedexTest#test_returns_gbp_instead_of_ukl_currency_for_uk_rates